requirements.txt
blinker==1.7.0
bottle==0.12.25
cffi==1.16.0
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
json2html==1.3.0
MarkupSafe==2.1.5
pillow==10.2.0
proxy_tools==0.1.0
pycparser==2.21
pythonnet==3.0.3
pywebview==5.0.5
typing_extensions==4.10.0
Werkzeug==3.0.1
python-dotenv>=0.19.2
authlib>=1.0